+++ package.json
@@ -20,6 +20,7 @@
     "react-redux": "^8.0.2",
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
+    "sass": "^1.54.4",
     "typescript": "^4.7.4",
     "web-vitals": "^2.1.4"
   },
