+++ src/App/components/SearchBox/SearchBox.module.scss
