+++ src/App/components/StartingPage/StartingPage.module.scss
