+++ src/App/components/User/Auth.tsx
@@ -3,7 +3,7 @@
 import React, { useState } from 'react';
 import { useDispatch } from 'react-redux';
 import Input from 'App/components/Reuse/Input';
-import classes from 'App/components/User/Auth.module.css';
+import classes from 'App/components/User/Auth.module.scss';
 import { authActions } from 'App/store/auth';
 import Button from 'App/components/Reuse/Button';
 import { useNavigate } from 'react-router-dom';
